Move WorksUtil into the internal_models library
As it doesn't actually contain any display-specific logic!
