+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/v1/DisplayLocationsV1SerialisationTest.scala
@@ -5,12 +5,14 @@ import uk.ac.wellcome.display.models.{
   DisplaySerialisationTestBase,
   WorksIncludes
 }
-import uk.ac.wellcome.display.test.util.{JsonMapperTestUtil, WorksUtil}
+import uk.ac.wellcome.display.test.util.JsonMapperTestUtil
 import uk.ac.wellcome.models.work.internal.{
   IdentifiedItem,
   IdentifiedWork,
   PhysicalLocation
 }
+import uk.ac.wellcome.models.work.internal._
+import uk.ac.wellcome.models.work.test.WorksUtil
 
 class DisplayLocationsV1SerialisationTest
     extends FunSpec
