+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/v1/DisplayPublicationDateV1SerialisationTest.scala
@@ -2,8 +2,9 @@ package uk.ac.wellcome.display.models.v1
 
 import org.scalatest.FunSpec
 import uk.ac.wellcome.display.models.DisplaySerialisationTestBase
-import uk.ac.wellcome.display.test.util.{JsonMapperTestUtil, WorksUtil}
+import uk.ac.wellcome.display.test.util.JsonMapperTestUtil
 import uk.ac.wellcome.models.work.internal.{IdentifiedWork, Period}
+import uk.ac.wellcome.models.work.test.WorksUtil
 
 class DisplayPublicationDateV1SerialisationTest
     extends FunSpec
