+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/v2/DisplayWorkV2SerialisationTest.scala
@@ -5,8 +5,9 @@ import uk.ac.wellcome.display.models.{
   DisplaySerialisationTestBase,
   WorksIncludes
 }
-import uk.ac.wellcome.display.test.util.{JsonMapperTestUtil, WorksUtil}
+import uk.ac.wellcome.display.test.util.JsonMapperTestUtil
 import uk.ac.wellcome.models.work.internal._
+import uk.ac.wellcome.models.work.test.WorksUtil
 
 class DisplayWorkV2SerialisationTest
     extends FunSpec
